Fix retrieval of latest version
If the last item on the list in the website is not actually the current 
latest version, then the previous method would fail. For example, as of 
10/18/2018 the last item on the website is a folder named "revoked". 
This folder was previously broken.

To fix the issue, I sort the href entries by their date, grep only lines 
that have valid hashes, then `tail -1` the results. This should work for 
the foreseable future.